Golden_Standard/pyproject.toml
[project]
name = "uni-ai-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.10.0",
    "langchain>=0.3.23",
    "langchain-community>=0.3.21",
    "langchain-mistralai>=0.2.10",
    "huggingface-hub>=0.30.1",
    "python-telegram-bot>=20.6",
    "supabase>=0.2.0",
    "python-dotenv>=0.21.0",
    "PyPDF2>=3.0.1",
]
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
